requirements.txt
# Core
python-dotenv>=1.0.0

# Data
faker>=24.0.0
numpy>=1.26.0
pandas>=2.1.0

# Date handling
python-dateutil>=2.9.0

# Scraping (optional)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Logging
loguru>=0.7.0
